requirements.txt
Django
pytz
pillow
pymysql
mysqlclient
django-tinymce
sqlparse
